requirements.txt
numpy
matplotlib
pillow
scikit-learn
pytorch
torchvision
tqdm
requests
opencv-python
jupyter
ipympl
